+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Import logging will now include the final list of transactions imported
 - Completely revamped CLI display for import and settlement info
 - `--verbose` flag can be used to display final imported transactions for imports
+- Added support for renaming tickers by managing ticker aliases with `folio tickers`
 
 ### Changed